virtual_tryon/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
torch>=1.7.0
torchvision>=0.8.1
opencv-python
scikit-image
